+++ NEWS.md
@@ -6,6 +6,10 @@
 * Fix bug in file size estimation when 'whole.recs = TRUE' in `selection_table()` 
 * Box and axis can e removed in `phylo_spectro()`
 
+## New functions:
+
+* `by_element_est()` convert a by-song extended selection table to by-element
+
 # *warbleR 1.1.27*
 
 ## New functions: